LayoutTests/fast/css/shadow-parts/exportparts-syntax-expected.txt

The following text should be green:

PASS Shortcut notation 
PASS Full notation 
PASS Skip bad entries 
PASS Mapping multiple parts to a single export 
FAIL Multiple exports for a single part assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 255)"
